Restructured the folders, and cleaned up code to not include unnecessary usings.
