+++ Src/LoLMatchAccepterNet.Common/NotificationServerContants.cs
