+++ Src/LoLMatchAccepterNet/GlobalSuppressions.cs
