+++ Src/LoLMatchAccepterNet/LCU/Game.cs
@@ -1,11 +1,4 @@
-﻿using LoLMatchAccepterNet.Api;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
-using System.Text;
-using System.Text.Json;
-using System.Threading.Tasks;
+﻿using System.Text.Json;
 
 namespace LoLMatchAccepterNet.LCU
 {
