+++ Src/LoLMatchAccepterNet/LCU/LcuClient.cs
@@ -1,13 +1,7 @@
-using LoLMatchAccepterNet.Api;
 using LoLMatchAccepterNet.LCU;
-using System;
 using System.Diagnostics;
-using System.Net.Http;
 using System.Text;
-using System.Text.Json;
 using System.Text.RegularExpressions;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace LeagueMatchAccepter
 {
