+++ Src/LoLMatchAccepterNet/LoLMatchAccepterNet.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\LoLMatchAccepter.Common\LoLMatchAccepter.Common.csproj" />
+    <ProjectReference Include="..\LoLMatchAccepterNet.Common\LoLMatchAccepterNet.Common.csproj" />
   </ItemGroup>
 
 </Project>
