+++ Src/LoLMatchNotificator/App.xaml
