+++ Src/LoLMatchNotificator/App.xaml.cs
