+++ Src/LoLMatchNotificator/AppShell.xaml
