+++ Src/LoLMatchNotificator/AppShell.xaml.cs
