+++ Src/LoLMatchNotificator/Platforms/Android/AndroidManifest.xml
