+++ Src/LoLMatchNotificator/Platforms/Android/MainActivity.cs
@@ -1,11 +1,9 @@
 ﻿using Android.App;
-using Android.Content;
 using Android.Content.PM;
 using Android.OS;
-using MatchNotificator.Platforms.Android;
-using MatchNotificator.Services;
+using MatchNotificator.Platforms.Android.Permissions;
 
-namespace MatchNotificator
+namespace MatchNotificator.Platforms.Android
 {
     [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, LaunchMode = LaunchMode.SingleTop, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
     public class MainActivity : MauiAppCompatActivity
@@ -13,8 +11,7 @@ namespace MatchNotificator
         protected override async void OnCreate(Bundle? savedInstanceState)
         {
             base.OnCreate(savedInstanceState);
-
-            PermissionStatus status = await Permissions.RequestAsync<NotificationPermission>();
+            _ = await Microsoft.Maui.ApplicationModel.Permissions.RequestAsync<NotificationPermission>();
         }
     }
 }
