+++ Src/LoLMatchNotificator/Platforms/Android/Resources/values/colors.xml
