+++ Src/LoLMatchNotificator/Platforms/MacCatalyst/AppDelegate.cs
