+++ Src/LoLMatchNotificator/Platforms/MacCatalyst/Entitlements.plist
