+++ Src/LoLMatchNotificator/Platforms/MacCatalyst/Info.plist
