+++ Src/LoLMatchNotificator/Platforms/MacCatalyst/Program.cs
