+++ Src/LoLMatchNotificator/Platforms/Tizen/Main.cs
