+++ Src/LoLMatchNotificator/Platforms/Tizen/tizen-manifest.xml
