+++ Src/LoLMatchNotificator/Platforms/Windows/App.xaml.cs
