+++ Src/LoLMatchNotificator/Platforms/Windows/Package.appxmanifest
