+++ Src/LoLMatchNotificator/Platforms/Windows/app.manifest
