+++ Src/LoLMatchNotificator/Platforms/iOS/AppDelegate.cs
