+++ Src/LoLMatchNotificator/Platforms/iOS/Info.plist
