+++ Src/LoLMatchNotificator/Platforms/iOS/Program.cs
