+++ Src/LoLMatchNotificator/Platforms/iOS/Resources/PrivacyInfo.xcprivacy
