+++ Src/LoLMatchNotificator/Platforms/iOS/Services/NotificationService.cs
