+++ Src/LoLMatchNotificator/Properties/launchSettings.json
