+++ Src/LoLMatchNotificator/Resources/Fonts/OpenSans-Regular.ttf
