+++ Src/LoLMatchNotificator/Resources/Fonts/OpenSans-Semibold.ttf
