+++ Src/LoLMatchNotificator/Resources/Styles/Colors.xaml
