+++ Src/LoLMatchNotificator/Resources/Styles/Styles.xaml
