+++ Src/LoLMatchNotificator/Services/INotificationService.cs
@@ -1,0 +1,7 @@
+﻿namespace MatchNotificator.Services
+{
+    public interface INotificationService
+    {
+        public void NotifyUser();
+    }
+}
